Overhaul the application structure using prototypal inheritance
First up, the name of the application has been changed to Datacraft.
Datacraft is a JavaScript object that has useful features as prototype
properties. Defining the prototype, modules were created in order to split
the code base systematically.

'Processor', previously, has become 'job's and 'task's. A job is comprised
of one or more tasks which process files. Tasks can either read the data of
files or output of previous tasks as input.

Job and task each have configuration which is optional.